Bug 2104402: Merge wizard and vm Environment editors
