+++ src/utils/components/EnvironmentEditor/components/EnvironmentFormActions.scss
@@ -1,4 +1,4 @@
-.vm-environment-tab__buttons {
+.environment-form__buttons {
   border-top: 1px solid #ddd;
   padding-top: var(--pf-global--spacer--lg);
