+++ src/utils/components/EnvironmentEditor/hooks/useEnvironmentsResources.ts
