+++ src/views/catalog/wizard/tabs/environment/WizardEnvironmentTab.tsx
@@ -1,107 +1,21 @@
 import * as React from 'react';
-import { useParams } from 'react-router-dom';
 
 import { WizardTab } from '@catalog/wizard/tabs';
-import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
-import { Button } from '@patternfly/react-core';
-import { PlusCircleIcon } from '@patternfly/react-icons';
-
-import EnvironmentEditor from './components/EnvironmentEditor';
-import WizardEnvironmentTabFooter from './components/WizardEnvironmentTabFooter';
-import WizardEnvironmentTabSkeleton from './components/WizardEnvironmentTabSkeleton';
-import WizardEnvironmentTabTitle from './components/WizardEnvironmentTabTitle';
-import useEnvironments from './hook/useEnvironments';
-import useEnvironmentsResources from './hook/useEnvironmentsResources';
+import EnvironmentForm from '@kubevirt-utils/components/EnvironmentEditor/EnvironmentForm';
+import Loading from '@kubevirt-utils/components/Loading/Loading';
+import { Bullseye } from '@patternfly/react-core';
 
 const WizardEnvironmentTab: WizardTab = ({ vm, updateVM, setDisableVmCreate }) => {
-  const { ns } = useParams<{ ns: string }>();
-  const { t } = useKubevirtTranslation();
-
-  const {
-    secrets,
-    configMaps,
-    serviceAccounts,
-    loaded,
-    error: loadError,
-  } = useEnvironmentsResources(ns);
-
-  const {
-    environments,
-    onEnvironmentAdd,
-    onEnvironmentChange,
-    onEnvironmentRemove,
-    onReload,
-    onSave,
-    edited,
-    error: formError,
-    setError,
-  } = useEnvironments(vm, updateVM);
-
-  React.useEffect(() => {
-    setDisableVmCreate(edited);
-    return () => setDisableVmCreate(false);
-  }, [edited, setDisableVmCreate]);
-
-  if (!loaded) return <WizardEnvironmentTabSkeleton />;
-
-  const environmentNamesSelected = environments.map((env) => env.name);
-
-  const closeError = () => {
-    setError(undefined);
-  };
+  if (!vm)
+    return (
+      <Bullseye>
+        <Loading />
+      </Bullseye>
+    );
 
   return (
     <div className="co-m-pane__body">
-      <WizardEnvironmentTabTitle />
-
-      {environments.length !== 0 && (
-        <div className="row pairs-list__heading">
-          <div className="col-xs-5 text-secondary text-uppercase" id="environment-name-header">
-            {t('config map / secret / service account')}
-          </div>
-          <div className="col-xs-5 text-secondary text-uppercase" id="environment-serial-header">
-            {t('Serial Number')}
-          </div>
-          <div className="col-xs-1 co-empty__header" />
-        </div>
-      )}
-
-      {environments.map((environment, index) => (
-        <EnvironmentEditor
-          key={environment.name}
-          environmentName={environment.name}
-          serial={environment?.serial}
-          kind={environment.kind}
-          secrets={secrets}
-          configMaps={configMaps}
-          serviceAccounts={serviceAccounts}
-          onChange={onEnvironmentChange}
-          onRemove={onEnvironmentRemove}
-          id={index}
-          environmentNamesSelected={environmentNamesSelected}
-        />
-      ))}
-
-      <div className="row">
-        <div className="col-xs-12">
-          <Button
-            className="pf-m-link--align-left"
-            onClick={onEnvironmentAdd}
-            type="button"
-            variant="link"
-          >
-            <PlusCircleIcon /> {t('Add Config Map, Secret or Service Account')}
-          </Button>
-        </div>
-      </div>
-
-      <WizardEnvironmentTabFooter
-        error={loadError || formError}
-        onSave={onSave}
-        onReload={onReload}
-        closeError={closeError}
-        isSaveDisabled={!edited || !environments.every((env) => env.name)}
-      />
+      <EnvironmentForm vm={vm} onEditChange={setDisableVmCreate} updateVM={updateVM} />
     </div>
   );
 };
