--- src/views/catalog/wizard/tabs/environment/components/WizardEnvironmentTabFooter.scss
@@ -1,8 +1,0 @@
-.wizard-environment-tab__buttons {
-  border-top: 1px solid #ddd;
-  padding-top: var(--pf-global--spacer--lg);
-
-  .pf-c-form {
-    margin-top: 0;
-  }
-}
