--- src/views/virtualmachines/details/tabs/environment/components/EnvironmentEditor.scss
@@ -1,30 +1,0 @@
-$drag-column-width: 28px;
-
-
-.co-empty__header {
-  padding: 0;
-  width: $drag-column-width;
-}
-
-
-.pairs-list__action {
-  padding: 0;
-  width: $drag-column-width;
-}
-
-
-.pairs-list__action-icon {
-  background: transparent;
-  border: 0;
-  display: flex;
-}
-
-.pairs-list__action-icon--reorder {
-  cursor: move;
-}
-
-
-.pairs-list__span-btns {
-  border: 0;
-  display: flex;
-}
