--- src/views/virtualmachines/details/tabs/environment/components/VirtualMachineEnvironmentTabTitle.scss
@@ -1,7 +1,0 @@
-.header-title-inline {
-  display: inline;
-}
-
-.icon-size-small {
-  font-size: var(--pf-global--FontSize--sm);
-}
