--- src/views/virtualmachines/details/tabs/environment/utils/constants.ts
@@ -1,17 +1,0 @@
-export enum EnvironmentKind {
-  secret = 'secret',
-  serviceAccount = 'serviceAccount',
-  configMap = 'configMap',
-}
-
-export type EnvironmentVariable = {
-  name: string;
-  serial: string;
-  kind: EnvironmentKind;
-};
-
-export const MapKindToAbbr = {
-  [EnvironmentKind.secret]: 'S',
-  [EnvironmentKind.configMap]: 'CM',
-  [EnvironmentKind.serviceAccount]: 'SA',
-};
